pyproject.toml
[tool.poetry]
name = "course-5-project"
version = "0.1.0"
description = ""
authors = ["Agata <agagor1961@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
mypy = "^1.11.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
